Redirect Return Users To Login
Proof of concept if we want to have an "About Us" index page.

If a user has logged in before (and the 'remember_me' and 'been_here'
cookies haven't expired), then redirect any requests on '/' (no path)
from 'index.html' to '/login'.
If the user is still logged in (either with a session, or via
'remember_me'), '/login' will redirect to '/home', skipping the loggin
since it is unnecessary.
The '/login' page contains a link to '/index.html' which if click will
not get redirected (since it is a path), and users can still see the
index page if they choose.