Make Nginx Optional
While configuring Nginx on linux for another project I realized that it
has a default install location, and while it can be moved, that requires
more work than on windows.

As such, there is now a new router file which replicates the same path
matching that Nginx uses. Now only Node and a DB is required to be up
and running.

While I was updating the Nginx README, I decided to go through all of
the README's and match their styling.