PM's In Theory Work
I was right, a hello fixed the socket.request.user undefined error
I was getting.

A room is created by user id, which can be used for PM's. I have tested
if the room exists successfully. The actual socket.on function remains
to be tested when a client side UI is in place.

Lobby's still remain untested.