Functional / Fully Featured Nav
We now use the brower's history, so it is possible to use the back
arrow. There is the potential to add in states down the line if needed.

Through magic the app will also recognize urls, so you can hit refresh,
or even use a bookmark (although given that we have not set urls, this
would be a poor choice).

This combined with the remember me functionality means that the app
knows where you want to go, and brings it up immediatly without any
login or nav needed on the clients part.